+++ src/main/scala/com/github/derghust/pekkotemplate/message/ArgumentMessage.scala
@@ -1,3 +1,0 @@
-package com.github.derghust.pekkotemplate.message
-
-final case class ArgumentMessage(cmd: String, args: List[String]) extends Message