+++ src/main/scala/com/github/derghust/rensa_no_hibiki/config/PasswordSupervisorConfig.scala
@@ -1,0 +1,3 @@
+package com.github.derghust.rensa_no_hibiki.config
+
+case class PasswordSupervisorConfig(numberOfWorker: Int)