+++ src/main/scala/com/github/derghust/rensa_no_hibiki/message/InitMessage.scala
@@ -1,0 +1,6 @@
+package com.github.derghust.rensa_no_hibiki.message
+
+import com.github.derghust.rensa_no_hibiki.util.TypeExtension.*
+
+trait InitMessage:
+  val arguments: Array[(String, Int |: String)]