+++ src/main/scala/com/github/derghust/rensa_no_hibiki/structure/Entity.scala
@@ -1,0 +1,5 @@
+package com.github.derghust.rensa_no_hibiki.structure
+
+import com.github.derghust.rensa_no_hibiki.structure.GenericType.Id
+
+final case class Entity(id: Id, label: String, description: String)